requirements.txt
python-dotenv
boto3
pylint
pytest
pandas
psycopg2-binary
confluent-kafka
json
